Add Bonsell text to About page
> Thought about this some more this weekend and added some text.

Our fieldwork takes place within the homeland of the Iñupiat, past and present. We honor the people who have acted as stewards of this land since time immemorial, and respect the significant traditional ecological knowledge that is passed down by elders.  
<p>
We engage with local communities in multiple ways. We hold regular, open meetings with the communities of  Kaktovik and Utqiaġvik, as well as with Traditional Knowledge holders in Kaktovik. This project also includes a strong commitment to K-12 Programs in Kaktovik and Utqiaġvik and offers employment opportunities to native high school seniors or recently graduated students living in these villages as field research assistants.
</p>